Add files via upload
21.1 11.52
I have added two functions. 
combine_duplicates takes those pre-consensus queries that didn't end up with unique values. It checks that they have identical barcodes, and if they do then it keeps one of them in a new df, matching_df. Once this df is completed it is then concatenated with the consensus_df. I still need to build in a check to make sure that all queries in the shared_matches have an entry in the consensus_df. It might also be good to have a check to make sure they are all unique too. These could both be helper functions I think. 
concatenate_dfs is a new helper function that can be called across the programs. It takes 2+ dfs and concatenates them after first checking that they have the same column headers and in the same order.