Add files via upload
22.1 10.51
New function added - check_unique_values.
It takes 2 dfs and compares the entries in a column passed to it. This returns a print statement providing information as to how the 2 dfs compare. 
The line 'check_unique_values(consensus_df, shared_matches_df, 'Query#')' in the add_majority_consensus_sequences function currently calls it within this program. The aim here is to check that all of the queries in the shared_matches_df  have been given a majority consensus sequence.