Add files via upload
22.1 11.33
New function added - check_df_column_unique_entries_only (df, column_header)
The function checks a df to see if there are any duplicated values in a column. It prints out a statement in either case, though lets the program go ahead.
I've also updated the check_unique_values function so that it prints out the name of the dfs being looked at. They need to be defined before the function is called.