Add files via upload
25.1 10.31
New program to collate regional mapping barcodes. The combined csv output can then be copy and pasted (values only) into the formatted Excel sheet.
This is set up for the V1V3 sheet and Full 16S. I just need to add the ability for the other regions to be incorporated now.