Add files via upload
25.1 11.02
Updated to allow multiple regions to be combined.
Added write_csvs function so that all of the regions get written to a .csv - they can be copied into the main .xls file as tabs. The 'combined.csv' sheet is to be copy and pasted (values only) into the main tab of the .xls file, after deleting the old values.