Add files via upload
1.2 10.09
This is a program that reads in the output of euclidean_scoring.py and produces .csv files that can be used to produce graphs. 
The program also reformats the euclidean input and the output .csv file produced can be (and is currently) used to paste values on top of the old sheet.